Phase 1 start: vision.py (frozen SigLIP2) + modal_mm.py (download + vision smoke)
- vision.py: SiglipVision wrapper (google/siglip2-so400m-patch14-384), frozen, on-the-fly
  encode (precomputing 558K feats would be ~900GB). Lazy transformers import.
- modal_mm.py: moe-vlm Modal app; download LLaVA-Pretrain (LAION-CC-SBU-558K); GPU smoke =
  real SigLIP2 + 500M_ctx2048 backbone + MoEVLM fwd/bwd on synthetic image.
- multimodal.py: cast spliced features to text-embed dtype (autocast bf16/fp32 cat fix).
- test_vlm.py: +autocast dtype test (6/6 pass).

+++ multimodal.py
@@ -84,7 +84,7 @@ class MoEVLM(nn.Module):
             e_parts, t_parts, prev = [], [], 0
             for pos, feat in spots:
                 e_parts.append(text_emb[prev:pos])
-                e_parts.append(feat)
+                e_parts.append(feat.to(text_emb.dtype))   # encoder/projector may be bf16 under autocast
                 if targets is not None:
                     t_parts.append(targets[b][prev:pos])
                     t_parts.append(torch.full((feat.shape[0],), IGNORE_INDEX, dtype=targets.dtype, device=dev))
